# #32536 Speedup test_car_interfaces.py 