@sshane 2024-05-28
Closing since this strictly decreases test coverage when we want to increase it. There are likely optimization to be had in the car interface itself, as that's a big part of this test